Initial commit of 3 uploader - adapters.
You all know IFrame & Flash adapters form FancyUpload and there's the new one based on HTML5 XHR send file ability. 

Current implementation is aimed soloely on uploading large files without reading it in memory. This is why the request is not really valid multibyte submission. The version that does proper request (and thus does base64 conversion on the client side) may be implemented in future, but it's not useful for me (I need to support really large files).

Other things 
- Cleaned up for 1.3.
- Uploader.Task is a base class for all tasks of all adapters
- Untested yet. Sorry. A few days later.